pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "truthscore-llm"
version = "0.2.0"
description = "Claim-grounded truth evaluation for LLM outputs: retrieval, per-claim verification, and explicit risk signals"
readme = "README.md"
requires-python = ">=3.8"
dependencies = []
authors = [
    {name = "TruthScore Contributors"}
]
keywords = ["llm", "truth", "evaluation", "evidence", "consistency", "research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
nli = [
    "transformers>=4.0.0",
    "torch>=1.9.0",
]
judge = [
    "openai>=1.0.0",
]
retrieval = [
    "faiss-cpu>=1.7.4",
    "numpy>=1.20.0",
    "sentence-transformers>=2.2.0",
]

[project.urls]
Homepage = "https://github.com/mmsa/truthscore-llm"
Documentation = "https://github.com/mmsa/truthscore-llm#readme"
Repository = "https://github.com/mmsa/truthscore-llm"
Issues = "https://github.com/mmsa/truthscore-llm/issues"

[tool.setuptools]
packages = ["truthscore"]

[tool.setuptools.package-data]
truthscore = ["py.typed"]

